Split large components and consolidate frontend props
assignment-modal.tsx splits into the modal, the two confirm dialogs,
ProjectResultList with the suggestion empty state, and the pure helpers
in assignment-modal-logic.ts.
CalendarSection moves to its own file and the ZEP calendar discovery
state lives in a useZepCalendars hook, shrinking app.tsx and reducing
EmployeeIcalDialog to five props.
PlanningGrid and PlanningGridTable merge into one component whose
optional state props remain the test seam.
TimetableRow loses the duplicate onRetry callback (both callers passed
reloadAssignments) and groups the week data into one prop.
