+++ src/app.tsx
@@ -4,21 +4,15 @@ import { ChevronLeft, ChevronRight, Settings } from "lucide-react";
 import { EmployeeIcalDialog } from "./app/components/employee-ical-dialog";
 import { SettingsDialog } from "./app/components/settings/settings-dialog";
 import { usePlanningAssignments } from "./app/hooks/use-planning-assignments";
+import { useZepCalendars } from "./app/hooks/use-zep-calendars";
 import { PlanningGrid } from "./app/page";
 import { getWeekDays } from "./app/util";
-import type {
-  EmployeeSetting,
-  PlanningContactRecord,
-  ZepCalendar,
-} from "./generated/tauri";
+import type { EmployeeSetting, PlanningContactRecord } from "./generated/tauri";
 import { commands } from "./generated/tauri";
 import { loadDayliteContacts } from "./services/daylite-contacts";
-import { discoverZepCalendars } from "./services/zep";
 
 function App() {
   const [weekOffset, setWeekOffset] = useState(0);
-  // Display preference: show Saturday and Sunday in the planning view.
-  // Defaults to false (matches the backend DisplaySettings default).
   const [showWeekend, setShowWeekend] = useState(false);
   const weekStart = getWeekDays(weekOffset, showWeekend)[0]
     .toISOString()
@@ -34,15 +28,9 @@ function App() {
   const [employeeSettingsError, setEmployeeSettingsError] = useState<
     string | null
   >(null);
-  // Display preference: hide employees without a calendar or with category "Test".
-  // Defaults to true (matches the backend DisplaySettings default).
   const [hideNonPlannableEmployees, setHideNonPlannableEmployees] =
     useState(true);
-  // Session cache for discovered ZEP calendars (task 2.4). Not persisted across restarts;
-  // null means "not yet fetched", [] means "fetched but empty".
-  const [zepCalendars, setZepCalendars] = useState<ZepCalendar[] | null>(null);
-  const [isLoadingCalendars, setIsLoadingCalendars] = useState(false);
-  const [calendarsError, setCalendarsError] = useState<string | null>(null);
+  const zepCalendarsState = useZepCalendars();
 
   const loadEmployeeSettings = useCallback(async () => {
     const result = await commands.loadLocalStore();
@@ -89,31 +77,12 @@ function App() {
     };
   }, [loadEmployeeSettings]);
 
-  const loadZepCalendars = useCallback(async () => {
-    setIsLoadingCalendars(true);
-    setCalendarsError(null);
-    try {
-      const calendars = await discoverZepCalendars();
-      setZepCalendars(calendars);
-    } catch (error) {
-      setCalendarsError(
-        error instanceof Error
-          ? error.message
-          : "Die ZEP-Kalender konnten nicht geladen werden.",
-      );
-    } finally {
-      setIsLoadingCalendars(false);
-    }
-  }, []);
-
   const handleOpenIcalDialog = useCallback(
     (employee: PlanningContactRecord) => {
       setIcalDialogEmployee(employee);
-      if (zepCalendars === null && !isLoadingCalendars) {
-        void loadZepCalendars();
-      }
+      zepCalendarsState.ensureLoaded();
     },
-    [zepCalendars, isLoadingCalendars, loadZepCalendars],
+    [zepCalendarsState.ensureLoaded],
   );
 
   const handleIcalDialogClose = () => {
@@ -202,10 +171,7 @@ function App() {
         }
         onClose={handleIcalDialogClose}
         onSettingsSaved={handleSettingsSaved}
-        zepCalendars={zepCalendars}
-        isLoadingCalendars={isLoadingCalendars}
-        calendarsError={calendarsError}
-        onReloadCalendars={loadZepCalendars}
+        calendarState={zepCalendarsState}
       />
     </article>
   );
