+++ src/app/components/employee-ical-dialog.tsx
@@ -3,21 +3,25 @@ import type {
   EmployeeSetting,
   IcalSource,
   PlanningContactRecord,
-  ZepCalendar,
 } from "../../generated/tauri";
 import { saveAndTestCalendar } from "../../services/zep";
+import type { ZepCalendarsState } from "../hooks/use-zep-calendars";
+import { CalendarSection, type SectionStatus } from "./calendar-section";
 
 export function EmployeeIcalDialog({
   employee,
   employeeSetting,
   onClose,
   onSettingsSaved,
-  zepCalendars,
-  isLoadingCalendars,
-  calendarsError,
-  onReloadCalendars,
+  calendarState,
 }: Props) {
   const isOpen = employee !== null;
+  const {
+    calendars,
+    isLoading: isLoadingCalendars,
+    errorMessage: calendarsError,
+    reload: onReloadCalendars,
+  } = calendarState;
 
   const [primaryUrl, setPrimaryUrl] = useState("");
   const [absenceUrl, setAbsenceUrl] = useState("");
@@ -123,7 +127,7 @@ export function EmployeeIcalDialog({
           <CalendarSection
             title="Einsatz"
             source="primary"
-            calendars={zepCalendars}
+            calendars={calendars}
             selectedUrl={primaryUrl}
             storedUrl={employeeSetting?.zepPrimaryCalendar ?? ""}
             onUrlChange={(url) => {
@@ -133,13 +137,13 @@ export function EmployeeIcalDialog({
             status={primaryStatus ?? storedPrimaryStatus}
             isSubmitting={isPrimarySubmitting}
             onSubmit={() => void handleSubmit("primary")}
-            isDisabled={isLoadingCalendars || zepCalendars === null}
+            isDisabled={isLoadingCalendars || calendars === null}
           />
 
           <CalendarSection
             title="Abwesenheit"
             source="absence"
-            calendars={zepCalendars}
+            calendars={calendars}
             selectedUrl={absenceUrl}
             storedUrl={employeeSetting?.zepAbsenceCalendar ?? ""}
             onUrlChange={(url) => {
@@ -149,14 +153,14 @@ export function EmployeeIcalDialog({
             status={absenceStatus ?? storedAbsenceStatus}
             isSubmitting={isAbsenceSubmitting}
             onSubmit={() => void handleSubmit("absence")}
-            isDisabled={isLoadingCalendars || zepCalendars === null}
+            isDisabled={isLoadingCalendars || calendars === null}
             isOptional
           />
         </section>
 
         <section className="mt-6 flex items-center justify-between gap-2">
           {calendarsError !== null ||
-          (!isLoadingCalendars && zepCalendars !== null) ? (
+          (!isLoadingCalendars && calendars !== null) ? (
             <button
               type="button"
               className="btn btn-ghost btn-sm"
@@ -186,87 +190,6 @@ export function EmployeeIcalDialog({
         Schließen
       </button>
     </dialog>
-  );
-}
-
-export function CalendarSection({
-  title,
-  calendars,
-  selectedUrl,
-  storedUrl,
-  onUrlChange,
-  status,
-  isSubmitting,
-  onSubmit,
-  isDisabled,
-  isOptional,
-}: CalendarSectionProps) {
-  const isClearing = !selectedUrl && !!storedUrl;
-  return (
-    <section>
-      <h3 className="font-semibold text-sm mb-2">
-        {title}
-        {isOptional ? (
-          <span className="font-normal text-base-content/60 ml-1">
-            (optional)
-          </span>
-        ) : null}
-      </h3>
-
-      <div className="flex items-end gap-2">
-        <label className="form-control flex-1">
-          <select
-            className="select select-bordered select-sm w-full"
-            value={selectedUrl}
-            onChange={(e) => onUrlChange(e.target.value)}
-            disabled={isDisabled || isSubmitting}
-          >
-            <option value="">— Kein Kalender —</option>
-            {(calendars ?? []).map((cal) => (
-              <option key={cal.url} value={cal.url}>
-                {cal.displayName}
-              </option>
-            ))}
-          </select>
-        </label>
-        <button
-          type="button"
-          className={`btn btn-sm whitespace-nowrap ${isClearing ? "btn-error btn-outline" : "btn-primary"}`}
-          onClick={onSubmit}
-          disabled={isDisabled || isSubmitting || (!selectedUrl && !storedUrl)}
-        >
-          {isSubmitting
-            ? isClearing
-              ? "Entferne..."
-              : "Teste..."
-            : isClearing
-              ? "Entfernen"
-              : "Speichern & Testen"}
-        </button>
-      </div>
-
-      {isOptional && !selectedUrl && !status ? (
-        <p className="text-xs text-base-content/60 mt-1">
-          Ohne Abwesenheitskalender werden Abwesenheiten nicht synchronisiert.
-        </p>
-      ) : null}
-
-      {status ? (
-        <section
-          className={`alert alert-sm mt-2 py-2 ${
-            status.type === "success" ? "alert-success" : "alert-error"
-          }`}
-        >
-          <span className="text-sm">{status.message}</span>
-        </section>
-      ) : (
-        !isOptional &&
-        !isDisabled &&
-        !isSubmitting && (
-          <p className="text-xs text-base-content/60 mt-1">Nicht getestet.</p>
-        )
-      )}
-    </section>
   );
 }
 
@@ -344,27 +267,5 @@ interface Props {
   employeeSetting: EmployeeSetting | null;
   onClose: () => void;
   onSettingsSaved: () => void;
-  zepCalendars: ZepCalendar[] | null;
-  isLoadingCalendars: boolean;
-  calendarsError: string | null;
-  onReloadCalendars: () => void;
-}
-
-export interface CalendarSectionProps {
-  title: string;
-  source: IcalSource;
-  calendars: ZepCalendar[] | null;
-  selectedUrl: string;
-  storedUrl: string;
-  onUrlChange: (url: string) => void;
-  status: SectionStatus | null;
-  isSubmitting: boolean;
-  onSubmit: () => void;
-  isDisabled: boolean;
-  isOptional?: boolean;
-}
-
-export interface SectionStatus {
-  type: "success" | "error";
-  message: string;
+  calendarState: ZepCalendarsState;
 }
